+++ web/components/app-loaders/sandbox-app-loader.tsx
@@ -2,9 +2,7 @@ import BaseAppLoader from "@/components/app-loaders/base-app-loader";
 import Loading from "@/components/interface/status-screens/loading";
 import { EditorContext } from "@/components/providers/editor-context-provider";
 import { IMCContext } from "@/components/providers/imc-provider";
-import { PlatformEnum } from "@/lib/enums";
 import { usePlatformApi } from "@/lib/hooks/use-platform-api";
-import { getPlatform } from "@/lib/platform-api/platform-checker";
 import { ExtensionApp } from "@/lib/types";
 import { useDroppable } from "@dnd-kit/core";
 import {
@@ -178,7 +176,12 @@ export default function SandboxAppLoader({
         getHandlerMap(viewModel),
       );
     }
-  }, [viewModel, editorContext?.editorStates, editorContext?.persistSettings, platformApi]);
+  }, [
+    viewModel,
+    editorContext?.editorStates,
+    editorContext?.persistSettings,
+    platformApi,
+  ]);
 
   function getHandlerMap(model: ViewModel) {
     const newMap = new Map<IMCMessageTypeEnum, ReceiverHandler>();
@@ -214,87 +217,6 @@ export default function SandboxAppLoader({
       },
     );
 
-    // The following message handlers require OS-like environment.
-    // This can be either local environment or remote workspace.
-    newMap.set(
-      IMCMessageTypeEnum.PlatformWriteFile,
-      async (
-        senderWindow: Window,
-        message: IMCMessage,
-        abortSignal?: AbortSignal,
-      ) => {
-        if (message.payload) {
-          const { uri, file }: { uri: string; file: File | undefined } =
-            message.payload;
-
-          if (!file) {
-            throw new Error("File is undefined.");
-          }
-
-          const projectPath =
-            editorContext?.persistSettings?.projectHomePath +
-            "/" +
-            editorContext?.editorStates.project;
-
-          // Prevent writing to path outside the project path
-          if (!uri.startsWith(projectPath)) {
-            throw new Error(
-              "Cannot write to path outside the project directory.",
-            );
-          }
-          await platformApi?.writeFile(file, uri);
-        }
-      },
-    );
-    newMap.set(
-      IMCMessageTypeEnum.PlatformReadFile,
-      async (
-        senderWindow: Window,
-        message: IMCMessage,
-        abortSignal?: AbortSignal,
-      ) => {
-        const { uri }: { uri: string } = message.payload;
-
-        const projectPath =
-          editorContext?.persistSettings?.projectHomePath +
-          "/" +
-          editorContext?.editorStates.project;
-
-        // Prevent reading path outside the project path
-        if (!uri.startsWith(projectPath)) {
-          throw new Error(
-            `Cannot read file outside the project directory: ${uri}, project path: ${projectPath}`,
-          );
-        }
-
-        const file = await platformApi?.readFile(uri);
-        return file;
-      },
-    );
-    newMap.set(
-      IMCMessageTypeEnum.PlatformCreateTerminal,
-      async (
-        senderWindow: Window,
-        message: IMCMessage,
-        abortSignal?: AbortSignal,
-      ) => {
-        const platform = getPlatform();
-        // Get a shell terminal from native platform APIs
-        if (platform === PlatformEnum.Capacitor) {
-          return {
-            websocketUrl: editorContext?.persistSettings?.mobileHost,
-            projectHomePath: `~/storage/shared/${editorContext?.persistSettings?.projectHomePath}`,
-          };
-        } else {
-          const wsUrl = await platformApi?.createTerminal();
-          return {
-            websocketUrl: wsUrl,
-            projectHomePath: editorContext?.persistSettings?.projectHomePath,
-          };
-        }
-      },
-    );
-
     return newMap;
   }
 
@@ -314,10 +236,7 @@ export default function SandboxAppLoader({
   }
 
   return (
-    <div
-      className="relative h-full w-full"
-      ref={setNodeRef}
-    >
+    <div className="relative h-full w-full" ref={setNodeRef}>
       {isLookingForExtension ? (
         <div className="bg-content1 h-full w-full">
           <Loading />