+++ web/lib/types.ts
@@ -2,6 +2,7 @@ import {
   Action,
   Agent,
   AppConfig,
+  FileSystemObject,
   ModelConfig,
   PolyIMC,
   TTSModelConfig,
@@ -172,13 +173,6 @@ export type SaveFileDialogConfig = {
   extension?: string;
 };
 
-export type FileSystemObject = {
-  name: string;
-  uri: string;
-  isFolder: boolean;
-  subDirItems?: FileSystemObject[];
-};
-
 export type TreeViewGroupRef = {
   startCreatingNewFolder: () => void;
   startCreatingNewFile: () => void;
@@ -196,12 +190,6 @@ export type ContextMenuState = {
   x: number;
   y: number;
   isOpen: boolean;
-};
-
-export type ListPathOptions = {
-  include: "folders" | "files" | "all";
-  isRecursive?: boolean;
-  gitignore?: string[];
 };
 
 export type TabItem = {