+++ npm-packages/react-api/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @pulse-editor/react-api
 
+## 0.1.1-beta.62
+
+### Patch Changes
+
+- Update to react 19.2.1
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-beta.62
+
+## 0.1.1-beta.61
+
+### Patch Changes
+
+- Update packages and fix react versions
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-beta.61
+
 ## 0.1.1-beta.60
 
 ### Patch Changes